another way to fix IE11 draggables:
IE seems to need the default behavior for `dragenter` to be prevented in
order to get to the `dragover` event in some cases.

As mentioned in #9, there's another way to do this:
IE will respect the `dragover` event without any
issue if the element is something other than a basic `display: block`.

This pull request suggests another way to solve the problem.